Pull #11617: separates detail ast from xpath
