+++ src/main/java/com/puppycrawl/tools/checkstyle/gui/TreeTable.java
@@ -45,7 +45,7 @@ import javax.swing.tree.TreePath;
 
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
 import com.puppycrawl.tools.checkstyle.utils.XpathUtil;
-import com.puppycrawl.tools.checkstyle.xpath.AbstractNode;
+import com.puppycrawl.tools.checkstyle.xpath.ElementNode;
 import com.puppycrawl.tools.checkstyle.xpath.RootNode;
 import com.puppycrawl.tools.checkstyle.xpath.XpathQueryGenerator;
 import net.sf.saxon.trans.XPathException;
@@ -212,8 +212,8 @@ public final class TreeTable extends JTable {
                 final Deque<DetailAST> nodes =
                         XpathUtil.getXpathItems(xpath, new RootNode(rootAST))
                               .stream()
-                              .map(AbstractNode.class::cast)
-                              .map(AbstractNode::getUnderlyingNode)
+                              .map(ElementNode.class::cast)
+                              .map(ElementNode::getUnderlyingNode)
                               .collect(Collectors.toCollection(ArrayDeque::new));
                 updateTreeTable(xpath, nodes);
             }
