+++ src/main/java/com/puppycrawl/tools/checkstyle/xpath/RootNode.java
@@ -24,28 +24,11 @@ import java.util.List;
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
 import com.puppycrawl.tools.checkstyle.api.TokenTypes;
 import com.puppycrawl.tools.checkstyle.utils.XpathUtil;
-import com.puppycrawl.tools.checkstyle.xpath.iterators.DescendantIterator;
-import net.sf.saxon.Configuration;
-import net.sf.saxon.om.AxisInfo;
-import net.sf.saxon.om.GenericTreeInfo;
-import net.sf.saxon.om.NodeInfo;
-import net.sf.saxon.tree.iter.ArrayIterator;
-import net.sf.saxon.tree.iter.AxisIterator;
-import net.sf.saxon.tree.iter.EmptyIterator;
-import net.sf.saxon.tree.iter.SingleNodeIterator;
-import net.sf.saxon.type.Type;
 
 /**
- * Represents root node of Xpath-tree.
- *
+ * Represents DetailAST's root node of Xpath-tree.
  */
-public class RootNode extends AbstractNode {
-
-    /** Name of the root element. */
-    private static final String ROOT_NAME = "ROOT";
-
-    /** Constant for optimization. */
-    private static final AbstractNode[] EMPTY_ABSTRACT_NODE_ARRAY = new AbstractNode[0];
+public class RootNode extends AbstractRootNode {
 
     /** The ast node. */
     private final DetailAST detailAst;
@@ -56,20 +39,7 @@ public class RootNode extends AbstractNode {
      * @param detailAst reference to {@code DetailAST}
      */
     public RootNode(DetailAST detailAst) {
-        super(new GenericTreeInfo(Configuration.newConfiguration()));
         this.detailAst = detailAst;
-    }
-
-    /**
-     * Compares current object with specified for order.
-     * Throws {@code UnsupportedOperationException} because functionality not required here.
-     *
-     * @param nodeInfo another {@code NodeInfo} object
-     * @return number representing order of current object to specified one
-     */
-    @Override
-    public int compareOrder(NodeInfo nodeInfo) {
-        throw throwUnsupportedOperationException();
     }
 
     /**
@@ -91,109 +61,6 @@ public class RootNode extends AbstractNode {
     @Override
     public boolean hasChildNodes() {
         return detailAst != null;
-    }
-
-    /**
-     * Returns attribute value. Throws {@code UnsupportedOperationException} because root node
-     * has no attributes.
-     *
-     * @param namespace namespace
-     * @param localPart actual name of the attribute
-     * @return attribute value
-     */
-    @Override
-    public String getAttributeValue(String namespace, String localPart) {
-        throw throwUnsupportedOperationException();
-    }
-
-    /**
-     * Returns local part.
-     *
-     * @return local part
-     */
-    @Override
-    public String getLocalPart() {
-        return ROOT_NAME;
-    }
-
-    /**
-     * Returns type of the node.
-     *
-     * @return node kind
-     */
-    @Override
-    public int getNodeKind() {
-        return Type.DOCUMENT;
-    }
-
-    /**
-     * Returns parent.
-     *
-     * @return parent
-     */
-    @Override
-    public NodeInfo getParent() {
-        return null;
-    }
-
-    /**
-     * Returns root of the tree.
-     *
-     * @return root of the tree
-     */
-    @Override
-    public NodeInfo getRoot() {
-        return this;
-    }
-
-    /**
-     * Determines axis iteration algorithm. Throws {@code UnsupportedOperationException} in case,
-     * when there is no axis iterator for given axisNumber.
-     *
-     * @param axisNumber element from {@code AxisInfo}
-     * @return {@code AxisIterator} object
-     */
-    @Override
-    public AxisIterator iterateAxis(int axisNumber) {
-        final AxisIterator result;
-        switch (axisNumber) {
-            case AxisInfo.ANCESTOR:
-            case AxisInfo.ATTRIBUTE:
-            case AxisInfo.PARENT:
-            case AxisInfo.FOLLOWING:
-            case AxisInfo.FOLLOWING_SIBLING:
-            case AxisInfo.PRECEDING:
-            case AxisInfo.PRECEDING_SIBLING:
-                result = EmptyIterator.ofNodes();
-                break;
-            case AxisInfo.ANCESTOR_OR_SELF:
-            case AxisInfo.SELF:
-                result = SingleNodeIterator.makeIterator(this);
-                break;
-            case AxisInfo.CHILD:
-                if (hasChildNodes()) {
-                    result = new ArrayIterator.OfNodes<>(
-                            getChildren().toArray(EMPTY_ABSTRACT_NODE_ARRAY));
-                }
-                else {
-                    result = EmptyIterator.ofNodes();
-                }
-                break;
-            case AxisInfo.DESCENDANT:
-                if (hasChildNodes()) {
-                    result = new DescendantIterator(this, DescendantIterator.StartWith.CHILDREN);
-                }
-                else {
-                    result = EmptyIterator.ofNodes();
-                }
-                break;
-            case AxisInfo.DESCENDANT_OR_SELF:
-                result = new DescendantIterator(this, DescendantIterator.StartWith.CURRENT_NODE);
-                break;
-            default:
-                throw throwUnsupportedOperationException();
-        }
-        return result;
     }
 
     /**
@@ -234,25 +101,6 @@ public class RootNode extends AbstractNode {
     @Override
     public DetailAST getUnderlyingNode() {
         return detailAst;
-    }
-
-    /**
-     * Getter method for node depth.
-     *
-     * @return always {@code 0}
-     */
-    @Override
-    public int getDepth() {
-        return 0;
-    }
-
-    /**
-     * Returns UnsupportedOperationException exception.
-     *
-     * @return UnsupportedOperationException exception
-     */
-    private static UnsupportedOperationException throwUnsupportedOperationException() {
-        return new UnsupportedOperationException("Operation is not supported");
     }
 
 }
